requirements.txt
pymysql>=1.1.0
requests>=2.31.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
jinja2>=3.1.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
